Some old stuff I forgot to push

+++ graph/model/models_gen.go
@@ -39,13 +39,13 @@ type NewUserInput struct {
 }
 
 type Position struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 
 type PositionInput struct {
-	X *int `json:"x"`
-	Y *int `json:"y"`
+	X *float64 `json:"x"`
+	Y *float64 `json:"y"`
 }
 
 type TransitionInput struct {
